Upgrades OpenFst recipe to 1.7.5.
I am the upstream maintainer of OpenFst. This upgrades the recipe to
version 1.7.5 and enables all extensions. Since this also causes the
binaries to be installed we move tests to basic shell tests instead of
the compilation-based test in the older version.

This is my first conda-forge submission so please be kind ;).